Bookly/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
httpx
pytest
pytest-asyncio
pytest-cov
black
isort
mypy
python-multipart
jinja2
email-validator
passlib[bcrypt]
itsdangerous
celery
redis
python-jose